Split-Amp & XCorr adjustments
impoly() is seemingly broken in 2024b or fully depreciated. Split-Amp functionality was changed over to using drawpolygon() instead. That change required adjustment of the x-axis of the Amplitudes plot to be changed over to time indexing in seconds as a double rather than as a duration as drawpolygon() doesn't know how to handle the duration x-axis. Adjustments to the Amplitudes and individual trace plots were also made to account for this change.

XCorr adjustments included a conditional statement to enable units with 0-1 spikes in them to not break SortaSort. These instances arose from Split-Amp of either an accidental split with no spikes inside, or selection of a single noise spike of large amplitude. This change just skips over plotting instances where one of the groups has <2 spikes in it.